pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "picture-tool"
version = "0.2.0"
description = "Production-ready image pipeline toolkit: augmentation, splitting, YOLO training, CLI and GUI orchestration."
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "VisionSuite Team" }]
license = "LicenseRef-Proprietary"
keywords = ["yolo", "augmentation", "computer-vision", "training", "tooling"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Image Processing",
  "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]
dependencies = [
  "albumentations>=1.3",
  "matplotlib>=3.5",
  "numpy>=1.21",
  "opencv-python>=4.5",
  "PyYAML>=6.0",
  "scikit-learn>=1.0",
  "tqdm>=4.60",
  "ultralytics>=8.0",
]
license-files = ["LICENSE"]

[project.optional-dependencies]
gui = ["PyQt5>=5.15"]
dev = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "mypy>=1.11",
]

[project.scripts]
picture-tool-pipeline = "picture_tool.main_pipeline:main"
picture-tool-gui = "picture_tool.gui.app:main"

[project.urls]
Homepage = "https://github.com/robotlearning/Yolo11_auto_train"
Repository = "https://github.com/robotlearning/Yolo11_auto_train"
Issues = "https://github.com/robotlearning/Yolo11_auto_train/issues"
Documentation = "https://github.com/robotlearning/Yolo11_auto_train/tree/main/docs"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["picture_tool*"]

[tool.setuptools.package-data]
picture_tool = [
  "py.typed",
  "**/*.yaml",
]
